Test parsing duplex IO stream
I added this test when debugging Ox, which hangs when given a duplex
stream.

I managed to trick it to read the stream partially by monkey-patching
the IO object in the test. @ohler55, I’m wondering if this is something
that should be changed in Ox?

The other wrappers handle the stream fine.